+++ stroom-data/stroom-data-retention-impl/src/main/java/stroom/data/retention/impl/DataRetentionModule.java
@@ -49,17 +49,7 @@ public class DataRetentionModule extends AbstractModule {
     @SuppressWarnings("unused") // called by guice
     @Provides
     DataRetentionRules getRules(final DataRetentionRulesService dataRetentionRulesService) {
-        DataRetentionRules dataRetentionRules = null;
-        final Set<DocRef> set = dataRetentionRulesService.listDocuments();
-        if (set != null && set.size() == 1) {
-            dataRetentionRules = dataRetentionRulesService.readDocument(set.iterator().next());
-        }
-
-        if (dataRetentionRules != null) {
-            return dataRetentionRules;
-        }
-
-        return null;
+        return dataRetentionRulesService.getOrCreate();
     }
 
     private static class DataRetention extends RunnableWrapper {