+++ stroom-data/stroom-data-retention-impl/src/main/java/stroom/data/retention/impl/DataRetentionRulesService.java
@@ -21,4 +21,5 @@ import stroom.data.retention.shared.DataRetentionRules;
 import stroom.docstore.api.DocumentStore;
 
 public interface DataRetentionRulesService extends DocumentStore<DataRetentionRules> {
+    DataRetentionRules getOrCreate();
 }